+++ android-plugin/src/main/java/io/ecocode/java/JavaCheckList.java
@@ -29,7 +29,7 @@ import io.ecocode.java.checks.environment.optimized_api.FusedLocationRule;
 import io.ecocode.java.checks.environment.power.ChargeAwarenessRule;
 import io.ecocode.java.checks.environment.power.SaveModeAwarenessRule;
 import io.ecocode.java.checks.environment.sobriety.*;
-import io.ecocode.java.checks.social.privacy.GoogleAndroidAdsConsentRule;
+import io.ecocode.java.checks.social.gdpr.GoogleAndroidAdsConsentRule;
 import io.ecocode.java.checks.social.privacy.GoogleTrackerRule;
 import io.ecocode.java.checks.social.privacy.TrackingIdRule;
 import org.sonar.plugins.java.api.JavaCheck;