+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- add rule SGDP001 google ads consent
+- Rule EC533 google ads consent
 
 ### Changed
 - SPRI004: Add rule to avoid use of Tracking Id using TelephonyManager#getDeviceId()